requirements.txt
numpy
pandas
matplotlib
seaborn
Pillow
chainlit
google-genai


